Sistema_Inteligencia_Drogas/requirements.txt
sglang[all]
qwen-agent[gui,rag,code_interpreter,mcp]
python-dotenv